+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added an option to disable use of double-buffering when downloading flash (#1030, #883)
 - Added a permissions system that allows the user to specify if a full chip erase is allowed (#918)
 - Added debug sequence for the nRF5340 that turns on the network core can unlock both cores by erasing them if that is permitted (#918)
 - Support for core registers `msp`, `psp` and `extra`, extra containing: